+++ modules/core/src/main/java/org/apache/ignite/internal/commandline/cache/argument/TracingConfigurationCommandArg.java
@@ -19,9 +19,8 @@ package org.apache.ignite.internal.commandline.cache.argument;
 
 import org.apache.ignite.internal.commandline.CommandList;
 import org.apache.ignite.internal.commandline.argument.CommandArg;
-import org.apache.ignite.internal.processors.tracing.Scope;
+import org.apache.ignite.spi.tracing.Scope;
 import org.apache.ignite.internal.processors.tracing.Span;
-
 import java.util.Arrays;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -48,7 +47,7 @@ public enum TracingConfigurationCommandArg implements CommandArg {
      * Set of {@link Scope} that defines which sub-traces will be included in given trace. In other words, if child's
      * span scope is equals to parent's scope or it belongs to the parent's span included scopes, then given child span
      * will be attached to the current trace, otherwise it'll be skipped. See {@link
-     * Span#isChainable(org.apache.ignite.internal.processors.tracing.Scope)} for more details.
+     * Span#isChainable(Scope)} for more details.
      */
     INCLUDED_SCOPES("--included-scopes");
