+++ modules/core/src/main/java/org/apache/ignite/internal/commandline/tracing/configuration/TracingConfigurationArguments.java
@@ -19,7 +19,7 @@ package org.apache.ignite.internal.commandline.tracing.configuration;
 
 import java.util.Collections;
 import java.util.Set;
-import org.apache.ignite.internal.processors.tracing.Scope;
+import org.apache.ignite.spi.tracing.Scope;
 import org.apache.ignite.internal.processors.tracing.Span;
 import org.apache.ignite.internal.visor.tracing.configuration.VisorTracingConfigurationItem;
 import org.jetbrains.annotations.NotNull;
@@ -55,7 +55,7 @@ import org.jetbrains.annotations.Nullable;
      *  In other words, if child's span scope is equals to parent's scope
      *  or it belongs to the parent's span included scopes, then given child span will be attached to the current trace,
      *  otherwise it'll be skipped.
-     *  See {@link Span#isChainable(org.apache.ignite.internal.processors.tracing.Scope)} for more details.
+     *  See {@link Span#isChainable(Scope)} for more details.
      */
     private TracingConfigurationArguments(
         TracingConfigurationSubcommand cmd,
@@ -153,7 +153,7 @@ import org.jetbrains.annotations.Nullable;
          * In other words, if child's span scope is equals to parent's scope
          * or it belongs to the parent's span included scopes, then given child span will be attached to the current trace,
          * otherwise it'll be skipped.
-         * See {@link Span#isChainable(org.apache.ignite.internal.processors.tracing.Scope)} for more details.
+         * See {@link Span#isChainable(Scope)} for more details.
          * @return Builder
          */
         @SuppressWarnings("UnusedReturnValue")
