+++ modules/core/src/main/java/org/apache/ignite/internal/processors/tracing/DeferredSpan.java
@@ -20,6 +20,8 @@ package org.apache.ignite.internal.processors.tracing;
 import java.util.Collections;
 import java.util.Set;
 import java.util.function.Supplier;
+import org.apache.ignite.spi.tracing.Scope;
+import org.apache.ignite.spi.tracing.SpanStatus;
 
 /**
  * Encapsulates concept of a deferred-initialized span. It's used to overcome OpenCensus span implementation, that starts
