+++ modules/core/src/main/java/org/apache/ignite/internal/processors/tracing/NoopTracing.java
@@ -18,7 +18,7 @@
 package org.apache.ignite.internal.processors.tracing;
 
 import org.apache.ignite.internal.processors.tracing.configuration.NoopTracingConfigurationManager;
-import org.apache.ignite.internal.processors.tracing.configuration.TracingConfigurationManager;
+import org.apache.ignite.spi.tracing.TracingConfigurationManager;
 import org.apache.ignite.internal.processors.tracing.messages.TraceableMessagesHandler;
 import org.apache.ignite.logger.NullLogger;
 import org.jetbrains.annotations.NotNull;
