+++ modules/core/src/main/java/org/apache/ignite/internal/processors/tracing/Span.java
@@ -19,6 +19,8 @@ package org.apache.ignite.internal.processors.tracing;
 
 import java.util.Set;
 import java.util.function.Supplier;
+import org.apache.ignite.spi.tracing.Scope;
+import org.apache.ignite.spi.tracing.SpanStatus;
 
 /**
  * Logical piece of a trace that represents a single operation.
