+++ modules/core/src/main/java/org/apache/ignite/internal/processors/tracing/Tracing.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.internal.processors.tracing;
 
-import org.apache.ignite.internal.processors.tracing.configuration.TracingConfigurationManager;
+import org.apache.ignite.spi.tracing.TracingConfigurationManager;
 import org.apache.ignite.internal.processors.tracing.messages.TraceableMessagesHandler;
 import org.jetbrains.annotations.NotNull;
