+++ modules/core/src/main/java/org/apache/ignite/spi/tracing/NoopSpiSpecificSpan.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.processors.tracing;
+package org.apache.ignite.spi.tracing;
 
 import java.util.Map;
