+++ modules/core/src/main/resources/META-INF/classnames.properties
@@ -1768,6 +1768,8 @@ org.apache.ignite.internal.processors.task.GridTaskWorker$3
 org.apache.ignite.internal.processors.task.GridTaskWorker$5
 org.apache.ignite.internal.processors.task.GridTaskWorker$State
 org.apache.ignite.internal.processors.timeout.GridTimeoutProcessor$2
+org.apache.ignite.spi.tracing.SpanStatus
+org.apache.ignite.internal.processors.tracing.messages.SpanContainer
 org.apache.ignite.internal.sql.SqlLexerTokenType
 org.apache.ignite.internal.sql.SqlParseException
 org.apache.ignite.internal.sql.SqlStrictParseException
